Extrai modulo de cards de finance
