+++ apps/web/package.json
@@ -45,6 +45,7 @@
       "@maskito/react": "catalog:ui",
       "@modules/account": "workspace:*",
       "@modules/agents": "workspace:*",
+      "@modules/cards": "workspace:*",
       "@modules/classification": "workspace:*",
       "@modules/finance": "workspace:*",
       "@modules/inbox": "workspace:*",
