+++ apps/web/src/integrations/orpc/router/index.ts
@@ -2,11 +2,11 @@ import * as accountRouter from "@modules/account/router/profile";
 import * as agentChatRouter from "@modules/agents/router/chat";
 import * as agentSettingsRouter from "@modules/account/router/agent-settings";
 import * as apiKeysRouter from "@modules/account/router/api-keys";
+import * as creditCardsRouter from "@modules/cards/router/credit-cards";
 import * as bankAccountsRouter from "@modules/finance/router/bank-accounts";
 import * as categoriesRouter from "@modules/classification/router/categories";
 import * as categoriesBulkRouter from "@modules/classification/router/categories-bulk";
 import * as cnpjRouter from "@modules/account/router/cnpj";
-import * as creditCardsRouter from "@modules/finance/router/credit-cards";
 import * as financialSettingsRouter from "@modules/account/router/financial-settings";
 import * as inboxRouter from "@modules/inbox/router/inbox";
 import * as reportsRouter from "@modules/insights/router/reports";
