+++ apps/web/tsconfig.json
@@ -23,6 +23,7 @@
          "@packages/ui/*": ["../../packages/ui/src/*"],
          "@modules/account/*": ["../../modules/account/src/*"],
          "@modules/agents/*": ["../../modules/agents/src/*"],
+         "@modules/cards/*": ["../../modules/cards/src/*"],
          "@modules/classification/*": ["../../modules/classification/src/*"],
          "@modules/finance/*": ["../../modules/finance/src/*"],
          "@modules/inbox/*": ["../../modules/inbox/src/*"]
