+++ modules/cards/tsconfig.json
@@ -1,0 +1,23 @@
+{
+   "extends": "@tooling/typescript/core.json",
+   "compilerOptions": {
+      "composite": false,
+      "declaration": false,
+      "declarationMap": false,
+      "noEmit": true,
+      "rootDir": "../..",
+      "jsx": "react-jsx",
+      "paths": {
+         "@core/database": ["../../core/database/src/index.ts"],
+         "@core/database/*": ["../../core/database/src/*"],
+         "@core/orpc/*": ["../../core/orpc/src/*"],
+         "@modules/cards/*": ["./src/*"]
+      }
+   },
+   "include": ["src"],
+   "references": [
+      {
+         "path": "../../core/database"
+      }
+   ]
+}
